Update README.md
The reliability and economic competitiveness of a thermal system hinge upon its safety and regulatory measures. During the stage of system design and analysis, researchers and engineers typically leverage extensive experimental efforts and investigate evolutionary models that represent the state-of-the-art understanding in their fields of specialization—also known as domain knowledge (DK)—to predict various safety limits.